PN-115 Walls now connect to iron bars
