+++ src/main/java/cn/nukkit/block/BlockGlassPane.java
@@ -1,14 +1,7 @@
 package cn.nukkit.block;
 
-import cn.nukkit.api.PowerNukkitOnly;
-import cn.nukkit.api.Since;
 import cn.nukkit.item.Item;
-import cn.nukkit.math.BlockFace;
 import cn.nukkit.utils.BlockColor;
-
-import java.util.EnumSet;
-import java.util.Iterator;
-import java.util.Set;
 
 /**
  * Created on 2015/12/6 by xtypr.
@@ -19,46 +12,6 @@ public class BlockGlassPane extends BlockThin {
     public BlockGlassPane() {
     }
     
-    @PowerNukkitOnly
-    @Since("1.3.0.0-PN")
-    public boolean isStraight() {
-        Set<BlockFace> connections = getConnections();
-        if (connections.size() != 2) {
-            return false;
-        }
-
-        Iterator<BlockFace> iterator = connections.iterator();
-        BlockFace a = iterator.next();
-        BlockFace b = iterator.next();
-        return a.getOpposite() == b;
-    }
-    
-    @PowerNukkitOnly
-    @Since("1.3.0.0-PN")
-    public Set<BlockFace> getConnections() {
-        EnumSet<BlockFace> connections = EnumSet.noneOf(BlockFace.class);
-        for (BlockFace blockFace : BlockFace.Plane.HORIZONTAL) {
-            if (isConnected(blockFace)) {
-                connections.add(blockFace);
-            }
-        }
-        return connections;
-    }
-    
-    @PowerNukkitOnly
-    @Since("1.3.0.0-PN")
-    public boolean isConnected(BlockFace face) {
-        Block side = getSideAtLayer(0, face);
-        switch (side.getId()) {
-            case COBBLE_WALL:
-            case GLASS_PANE:
-            case STAINED_GLASS_PANE:
-                return true;
-            default:
-                return side.isSolid();
-        }
-    }
-
     @Override
     public String getName() {
         return "Glass Pane";
