+++ src/main/java/cn/nukkit/block/BlockThin.java
@@ -1,8 +1,16 @@
 package cn.nukkit.block;
 
+import cn.nukkit.api.PowerNukkitDifference;
+import cn.nukkit.api.PowerNukkitOnly;
+import cn.nukkit.api.Since;
 import cn.nukkit.math.AxisAlignedBB;
+import cn.nukkit.math.BlockFace;
 import cn.nukkit.math.SimpleAxisAlignedBB;
 import cn.nukkit.utils.LevelException;
+
+import java.util.EnumSet;
+import java.util.Iterator;
+import java.util.Set;
 
 /**
  * Created on 2015/12/6 by xtypr.
@@ -49,8 +57,48 @@ public abstract class BlockThin extends BlockTransparent {
         );
     }
 
+    @PowerNukkitDifference(info = "Fixed connection logic for BE 1.16.0", since = "1.3.0.0-PN")
     public boolean canConnect(Block block) {
-        return block.isSolid() || block.getId() == this.getId() || block.getId() == GLASS_PANE || block.getId() == GLASS;
+        switch (block.getId()) {
+            case GLASS_PANE:
+            case STAINED_GLASS_PANE:
+            case IRON_BARS:
+            case COBBLE_WALL:
+                return true;
+            default:
+                return block.isSolid();
+        }
     }
 
+    @PowerNukkitOnly
+    @Since("1.3.0.0-PN")
+    public boolean isStraight() {
+        Set<BlockFace> connections = getConnections();
+        if (connections.size() != 2) {
+            return false;
+        }
+
+        Iterator<BlockFace> iterator = connections.iterator();
+        BlockFace a = iterator.next();
+        BlockFace b = iterator.next();
+        return a.getOpposite() == b;
+    }
+
+    @PowerNukkitOnly
+    @Since("1.3.0.0-PN")
+    public Set<BlockFace> getConnections() {
+        EnumSet<BlockFace> connections = EnumSet.noneOf(BlockFace.class);
+        for (BlockFace blockFace : BlockFace.Plane.HORIZONTAL) {
+            if (isConnected(blockFace)) {
+                connections.add(blockFace);
+            }
+        }
+        return connections;
+    }
+
+    @PowerNukkitOnly
+    @Since("1.3.0.0-PN")
+    public boolean isConnected(BlockFace face) {
+        return canConnect(getSideAtLayer(0, face));
+    }
 }
